% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifesimulatoR-package.R
\docType{package}
\name{lifesimulatoR-package}
\alias{lifesimulatoR}
\alias{lifesimulatoR-package}
\title{lifesimulatoR: Educational Origin-of-Life Simulations}
\description{
An educational package for exploring origin-of-life concepts,
including molecular evolution, mutation, selection, diversity metrics,
protocells, and autocatalytic networks.
}
\details{
Main workflow:

\enumerate{
\item create_prebiotic_pool()
\item molecule_fitness()
\item replicate_molecules()
\item mutate_population()
\item evolve_generation()
\item simulate_abiogenesis()
}

Alternative models:

\itemize{
\item protocell_simulation()
\item autocatalytic_network()
}

Analysis tools:

\itemize{
\item summarize_molecules()
\item shannon_entropy()
\item plot_simulation()
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NoushinN/lifesimulatoR}
  \item Report bugs at \url{https://github.com/NoushinN/lifesimulatoR/issues}
}

}
\author{
\strong{Maintainer}: Noushin Nabavi

Authors:
\itemize{
  \item Noushin Nabavi
}

}
\keyword{internal}
