% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolution.R
\name{mutate_population}
\alias{mutate_population}
\title{Mutate a Molecular Population}
\usage{
mutate_population(
  molecules,
  alphabet = c("A", "U", "G", "C"),
  mutation_rate = 0.01
)
}
\arguments{
\item{molecules}{Character vector of symbolic molecular sequences.}

\item{alphabet}{Character vector used to build molecular sequences.}

\item{mutation_rate}{Probability that a copied character mutates.}
}
\value{
A character vector of mutated molecular sequences.
}
\description{
Applies sequence mutation to every molecule in a population.
}
\examples{
mutate_population(c("AUGC", "UUUU"), mutation_rate = 0.1)
}
