% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolution.R
\name{mutate_sequence}
\alias{mutate_sequence}
\title{Mutate a Symbolic Molecular Sequence}
\usage{
mutate_sequence(sequence, alphabet, mutation_rate)
}
\arguments{
\item{sequence}{A single symbolic molecular sequence.}

\item{alphabet}{Character vector used to build molecular sequences.}

\item{mutation_rate}{Probability that a copied character mutates.}
}
\value{
A mutated molecular sequence.
}
\description{
Randomly changes characters in a molecule according to a mutation rate.
}
\examples{
mutate_sequence("AUGC", mutation_rate = 0.2)
}
