% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replication.R
\name{replicate_molecules}
\alias{replicate_molecules}
\title{Replicate a Molecular Population}
\usage{
replicate_molecules(
  molecules,
  n_molecules = length(molecules),
  selection_strength = 1
)
}
\arguments{
\item{molecules}{Character vector of symbolic molecular sequences.}

\item{n_molecules}{Number of molecules to return.}

\item{selection_strength}{Strength of fitness-based selection. Use 0 for neutral drift.}
}
\value{
A character vector of parent sequences before mutation.
}
\description{
Samples parent molecules according to fitness and creates the next generation.
}
\examples{
replicate_molecules(c("AUGC", "UUUU"), n_molecules = 5)
}
