% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{shannon_entropy}
\alias{shannon_entropy}
\title{Calculate Shannon Entropy}
\usage{
shannon_entropy(x)
}
\arguments{
\item{x}{Numeric vector of counts or abundances.}
}
\value{
Shannon entropy.
}
\description{
Calculates Shannon entropy from a vector of counts or abundances.
}
\examples{
shannon_entropy(c(10, 5, 1))
}
